+++ package.json
@@ -8,6 +8,7 @@
     "build": "webpack --progress --colors --env.production"
   },
   "dependencies": {
+    "awesome-typescript-loader": "^5.2.1",
     "axios": "^0.19.2",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^5.0.0",
@@ -17,18 +18,17 @@
     "jquery": "^3.5.1",
     "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.14.0",
+    "raw-loader": "^4.0.1",
     "sass-loader": "^8.0.2",
-    "ts-loader": "^7.0.4",
     "ts-node": "^8.10.1",
     "typescript": "^3.9.3",
     "ua-parser-js": "^0.7.21",
     "url-loader": "^4.1.0",
     "vue": "^2.6.11",
-    "vue-loader": "^15.9.2",
     "vue-material": "^1.0.0-beta-10.2",
     "vue-property-decorator": "^8.4.2",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.11",
+    "vuelidate": "^0.7.5",
     "webextension-polyfill-ts": "^0.15.0",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11"
@@ -41,8 +41,8 @@
     "@types/vuelidate": "^0.7.13",
     "@types/webpack": "^4.41.13",
     "tslint": "^6.1.2",
-    "tslint-sonarts": "^1.9.0",
-    "vuelidate": "^0.7.5"
+    "tslint-loader": "^3.5.4",
+    "tslint-sonarts": "^1.9.0"
   },
   "keywords": [
     "tasks",
