+++ src/components/options/options.component.scss
