+++ src/components/popup/popup.component.scss
