+++ src/platform/chrome/manifest.json
@@ -59,7 +59,7 @@
     "options_ui": {
         "page": "html/options.html"
     },
-    "content_security_policy": "script-src 'self' 'wasm-eval'; object-src 'self'",
+    "content_security_policy": "script-src 'self' 'wasm-eval' 'unsafe-eval'; object-src 'self'",
     "content_scripts" : [
         {
             "matches": [
