--- src/ts/types/json.d.ts
@@ -1,9 +1,0 @@
-declare module '*.json' {
-    const value: any;
-    export default value;
-}
-
-declare module 'json!*' {
-    const value: any;
-    export default value;
-}
