--- src/ts/types/vue.d.ts
@@ -1,17 +1,0 @@
-declare module '*.vue' {
-    import Vue from 'vue';
-    export default Vue;
-}
-//
-// declare module 'vue/types/options' {
-//     import Vue from 'vue';
-//
-//     interface ComponentOptions<V extends Vue> {
-//         style?: any;
-//     }
-// }
-//
-// declare module "*.scss" {
-//     const content: string;
-//     export default content;
-// }
