+++ src/ts/typings/shims-html.d.ts
@@ -1,0 +1,4 @@
+declare module '*.html' {
+    const content: string;
+    export default content;
+}
