+++ src/ts/typings/shims-json.d.ts
@@ -1,0 +1,4 @@
+declare module '*.json' {
+    const content: any;
+    export default content;
+}
