+++ src/ts/typings/shims-scss.d.ts
@@ -1,0 +1,4 @@
+declare module '*.scss' {
+    const content: string;
+    export default content;
+}
