+++ src/ts/typings/shims-vue.d.ts
@@ -1,0 +1,7 @@
+import Vue from 'vue';
+
+declare module 'vue/types/options' {
+    interface ComponentOptions<V extends Vue> {
+        style?: any;
+    }
+}
