+++ webpack.config.ts
@@ -1,6 +1,5 @@
-import {Configuration, DefinePlugin} from 'webpack';
+import {Configuration, DefinePlugin, RuleSetRule} from 'webpack';
 import {CleanWebpackPlugin} from 'clean-webpack-plugin';
-import {VueLoaderPlugin} from 'vue-loader';
 import {resolve} from 'path';
 import MiniCssExtractPlugin from 'mini-css-extract-plugin';
 import CopyWebpackPlugin from 'copy-webpack-plugin';
@@ -29,110 +28,117 @@ export default (env: any = {}): Configuration => {
     console.log('Production: ', production);
     console.log('Platform  : ', platform);
 
-    return {
-        mode   : production ? 'production' : 'development',
-        devtool: production ? false : 'inline-source-map',
-        entry  : {
-            client    : `${__dirname}/src/ts/client.ts`,
-            popup     : `${__dirname}/src/ts/popup.ts`,
-            options   : `${__dirname}/src/ts/options.ts`,
-            background: `${__dirname}/src/ts/background.ts`,
+    const rules: RuleSetRule[] = [
+        {
+            test    : /\.ts$/,
+            exclude : /node_modules/,
+            loader  : 'awesome-typescript-loader',
+            options :{
+                transpileOnly: production
+            }
         },
-        output : {
-            path      : `${__dirname}/dist/`,
-            filename  : 'js/[name].js'
+        {
+            test    : /\.html$/,
+            loader  : 'raw-loader',
+            exclude : [
+                `${__dirname}/src/platform/generic/html/popup.html`,
+                `${__dirname}/src/platform/generic/html/options.html`
+            ]
         },
-        resolve: {
-            modules   : ['node_modules', 'src'],
-            extensions: ['.ts', '.js', '.vue', '.json'],
-            alias     : resolveTsAliases({params: {platform}})
-        },
-        module : {
-            rules: [
+        {
+            test    : /\.scss$/,
+            use     : [
+                MiniCssExtractPlugin.loader,
+                'vue-style-loader',
+                'css-loader',
                 {
-                    test  : /\.vue$/,
-                    loader: 'vue-loader'
-                },
-                {
-                    test   : /\.ts$/,
-                    exclude: /node_modules/,
-                    use    : [
-                           {
-                               loader : 'ts-loader',
-                               options: {
-                                   appendTsSuffixTo: [/\.vue$/]
-                               }
-                           }
-                    ]
-                },
-                {
-                    test   : /\.(png|jpg|gif|svg|eot|ttf|woff|woff2)$/,
-                    loader : 'url-loader',
-                    options: {
-                        limit : 50000,
-                        name  : '/assets/[name].[ext]',
+                    loader  : 'sass-loader',
+                    options : {
+                        sassOptions: {
+                            outputStyle: production ? 'compressed' : undefined
+                        }
                     }
-                },
-                {
-                    test: /\.scss$/,
-                    use : [
-                        {
-                            loader: 'vue-style-loader'
-                        },
-                        {
-                            loader: MiniCssExtractPlugin.loader
-                        },
-                        {
-                            loader: 'css-loader'
-                        },
-                        {
-                            loader : 'sass-loader',
-                            options: {
-                                sassOptions: {
-                                    outputStyle: production ? 'compressed' : undefined
-                                }
-                            }
-                        }
-                    ]
-                },
-                {
-                    test: /\.css$/,
-                    use : [
-                        {
-                            loader: 'vue-style-loader'
-                        },
-                        {
-                            loader: MiniCssExtractPlugin.loader
-                        },
-                        {
-                            loader: 'css-loader'
-                        }
-                    ]
                 }
             ]
         },
-        plugins: [
-            new DefinePlugin(
-                {
-                    'process.env': {
-                        NODE_ENV    : production ? '"production"' : '"development"',
-                        APP_VERSION : `"${config.version}"`,
-                        APP_NAME    : `"${config.name}"`,
-                        APP_PLATFORM: `"${platform}"`,
-                        BUILD_TARGET: `"${platform}"`
-                    }
+        {
+            test    : /\.(png|jpg|gif|svg|eot|ttf|woff|woff2)$/,
+            loader  : 'url-loader',
+            options : {
+                limit : 50000,
+                name  : '/assets/[name].[ext]',
+            }
+        },
+    ];
+
+    const plugins = [
+        new DefinePlugin(
+            {
+                'process.env': {
+                    NODE_ENV    : production ? '"production"' : '"development"',
+                    APP_VERSION : `"${config.version}"`,
+                    APP_NAME    : `"${config.name}"`,
+                    APP_PLATFORM: `"${platform}"`,
+                    BUILD_TARGET: `"${platform}"`
                 }
-            ),
-            new VueLoaderPlugin(),
-            new CopyWebpackPlugin([`src/platform/generic`, `src/platform/${platform}`]),
-            new MiniCssExtractPlugin({filename: 'css/[name].css'}),
-            new CleanWebpackPlugin(
+            }
+        ),
+        new CopyWebpackPlugin(
+            [
                 {
-                    cleanStaleWebpackAssets     : false,
-                    cleanOnceBeforeBuildPatterns: ['**/*'],
-                    cleanAfterEveryBuildPatterns: ['scss']
+                    from        : `src/platform/generic`,
+                    ignore      : ['**/ts/**'],
+                },
+                {
+                    from        : `src/platform/${platform}`,
+                    ignore      : ['**/ts/**'],
                 }
-            ),
-        ]
+            ]
+        ),
+        new MiniCssExtractPlugin({filename: 'css/[name].css'}),
+        new CleanWebpackPlugin(
+            {
+                cleanStaleWebpackAssets     : false,
+                cleanOnceBeforeBuildPatterns: ['**/*'],
+                cleanAfterEveryBuildPatterns: ['scss']
+            }
+        )
+    ];
+
+    if (production) {
+    } else {
+        rules.push(
+            {
+                test    : /\.ts$/,
+                exclude : /node_modules/,
+                enforce : 'pre',
+                loader  : 'tslint-loader'
+            }
+        );
+    }
+
+    return {
+        mode    : production ? 'production' : 'development',
+        devtool : production ? false : 'source-map',
+        entry   : {
+            client          : `${__dirname}/src/ts/client.ts`,
+            popup           : `${__dirname}/src/ts/popup.ts`,
+            options         : `${__dirname}/src/ts/options.ts`,
+            background      : `${__dirname}/src/ts/background.ts`,
+        },
+        output  : {
+            path            : `${__dirname}/dist/`,
+            filename        : 'js/[name].js'
+        },
+        resolve : {
+            modules         : ['node_modules', 'src'],
+            extensions      : ['.ts', '.js', '.json'],
+            alias           : {
+                vue: 'vue/dist/vue.esm.js',
+                ...resolveTsAliases({params: {platform}})
+            }
+        },
+        module  : {rules},
+        plugins
     };
 };
